requirements.txt
Flask==3.0.2
pymongo==4.6.1
python-dotenv==0.21.1
flask-login==0.6.3
bcrypt==4.1.2
pandas==2.2.1
plotly==5.19.0 